test coverage

+++ handler/delete.go
@@ -1,6 +1,7 @@
 package handler
 
 import (
+	"fmt"
 	"net/http"
 
 	"github.com/gin-gonic/gin"
@@ -9,6 +10,10 @@ import (
 
 type DeleteUserRequest struct {
 	Name string `json:"name"`
+}
+
+type DeleteLinkRequest struct {
+	UUID string `json:"uuid"`
 }
 
 func (h *HandlerService) DropUsersTable(ctx *gin.Context) {
@@ -43,6 +48,11 @@ func (h *HandlerService) DeleteUser(ctx *gin.Context) {
 		return
 	}
 
+	if req.Name == "" {
+		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error missing required name"})
+		return
+	}
+
 	// delete from db
 	sql := "DELETE FROM users WHERE name = ?"
 	_, err := h.maria_repo.Exec(sql, req.Name)
@@ -53,5 +63,40 @@ func (h *HandlerService) DeleteUser(ctx *gin.Context) {
 	}
 
 	ctx.JSON(http.StatusOK, gin.H{"msg": "Successfully deleted user"})
+
+}
+
+func (h *HandlerService) DeleteLink(ctx *gin.Context) {
+	defer func() {
+		h.sugaredLogger.Desugar().Sync()
+	}()
+
+	var req DeleteLinkRequest
+	ctx.Header("Content-Type", "application/json")
+
+	fmt.Println("This is the ctx:", ctx)
+
+	if err := ctx.BindJSON((&req)); err != nil {
+		fmt.Println("We should not be here:", req)
+		h.sugaredLogger.Errorw("Error not a valid delete link request", zap.Error(err))
+		ctx.JSON(http.StatusInternalServerError, gin.H{"error": "Error not a valid delete link request"})
+		return
+	}
+
+	fmt.Println("This is the uuid from the request:", req)
+	if req.UUID == "" {
+		ctx.JSON(http.StatusBadRequest, gin.H{"error": "error missing required uuid"})
+		return
+	}
+	// delete from db
+	sql := "DELETE FROM links WHERE uuid = ?"
+	_, err := h.maria_repo.Exec(sql, req.UUID)
+	if err != nil {
+		h.sugaredLogger.Errorw("Error deleting link", zap.Any("error", err))
+		ctx.JSON(http.StatusInternalServerError, gin.H{"error": "failed to remove link"})
+		return
+	}
+
+	ctx.JSON(http.StatusOK, gin.H{"msg": "Successfully deleted link"})
 
 }
